+++ src/components/BuyList/BuyList.tsx
@@ -2,7 +2,6 @@ import { useCallback, useEffect, useState } from "react";
 import queryString from "query-string";
 import useSignIn from "hooks/useSignIn";
 import noComments from "assets/image/noComments.png";
-import { UsedBookBuyListResponse } from "modules/Slices/usedBookDetail/types";
 import TextField from "@mui/material/TextField";
 import Autocomplete, { AutocompleteChangeReason } from "@mui/material/Autocomplete";
 import Box from "@mui/material/Box";
@@ -23,13 +22,6 @@ import { ReviewListEmptyParagraph, ReviewListEmptyWrapper } from "../Reviews/Rev
 import { BuyListWrapper, FlexBox } from "./styles";
 import { BuyListResponse } from "./types";
 import { UserReviewCell, UserReviewHeader } from "../UserReview/styles";
-
-export interface buyConfirmSubmitParam {
-  orderId: string;
-}
-export interface BuyList {
-  pages: UsedBookBuyListResponse[];
-}
 
 const BuyList = () => {
   const { signIn, dispatch } = useSignIn();
