+++ src/components/Rating/types.ts
@@ -3,3 +3,8 @@ export interface HoverRatingProps {
   rating: number;
   handleChange: (event: React.SyntheticEvent<Element, Event>, value: number | null) => void;
 }
+
+export interface ReviewRankProps {
+  ReviewRank: number;
+  title: string;
+}
