+++ src/components/TextArea/types.ts
@@ -1,0 +1,9 @@
+export interface TextareaAutosizeParam {
+  isLoggedIn: boolean;
+  onChange: (event: any) => void;
+  checkAuth: () => boolean;
+  value: string;
+  limit: number;
+  height: number;
+  placeholder: string;
+}
