+++ src/components/UsedBookDetail/UsedBookContent/UsedBookArea.tsx
@@ -27,22 +27,7 @@ import {
   UsedBookDetailButton,
   UsedBookWrapper,
 } from "../style";
-
-export interface UsedBookAreaProps {
-  title: string;
-  price: number;
-  content: string;
-  view: number;
-  uploadDate: Date;
-  tags: string[];
-  likeCount: number;
-  replyCount: number;
-  usedBookId: number;
-  sellerId: number;
-  saleState: string;
-  bookState: string;
-  liked: boolean;
-}
+import { UsedBookAreaProps } from "./types";
 
 const UsedBookArea = ({
   title,
