+++ src/components/UsedBookDetail/UsedBookStoreInformation/types.ts
@@ -1,0 +1,7 @@
+import { FavoriteCategories } from "src/modules/Slices/usedBookDetail/types";
+
+export interface UsedStoreUserContentParam {
+  sellerName: string;
+  favoriteCategories: FavoriteCategories[];
+  totalSales: number;
+}
