+++ src/components/UsedBookDetail/UsedBookStoreInformationBottom.tsx
@@ -3,10 +3,6 @@ import { FlexWrapper } from "./style";
 import UsedBookInquiry from "./UsedBookInquiry/UsedBookInquiry";
 import UsedBookStoreInformationArea from "./UsedBookStoreInformation/UsedBookStoreInformationArea";
 
-export interface UsedBookStoreInformationBottomParam {
-  usedBookId: string;
-}
-
 export const UsedBookStoreInformationBottom = () => {
   return (
     <>
